requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.15.0
openpyxl>=3.1.0
google-genai>=1.0.0
python-dotenv>=1.0.0
